+++ ci4/app/Config/Autoload.php
@@ -99,4 +99,5 @@ class Autoload extends AutoloadConfig
      */
     public $helpers = [
     ];
+
 }