+++ CHANGELOG.md
@@ -25,6 +25,7 @@
     * Features: 
       * Add Comment with url to GitHub commit
       * Update fields based on conditions. For example change status from "development" to "test"
+* Authenticated flow is changed from implicit to authorization flow with PKCE [Issue 30](https://github.com/4spacesdk/kubernetes-service-orchestrator/issues/30)
 
 ### Breaking changes
 * Google Artifact Registry integration credentials has been moved from environment variables to container image properties